Refactor interruptible-gnani.py to streamline imports, update language settings for Gnani STT.
